requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pypdf==5.1.0

# Gemini + env
google-genai==0.3.0
python-dotenv==1.0.1

# Optional but nice to have
orjson==3.10.11
